requirements.txt
numpy>=1.20.0
matplotlib>=3.4.0
opencv-python>=4.5.0
scikit-image>=0.18.0
Pillow>=8.0.0
scipy>=1.6.0
# tkinter is part of the standard library, not a PyPI package
# For PIL.ImageTk, install system package: sudo apt-get install python3-pil.imagetk
